fixed bug with options; moved RKAdaptiveStepsizeODESolver to rk_common; fixed bug with adjoint_params.
